Remove unnecessary second allocation in GC Box
Previously, the value contained in Box<T> was stored as an allocated
pointer. This required two allocations for each object managed by the GC
- one for the Box and one for the object itself.
With this change, contained values are now stored in a byte buffer
defined as Box's member variable. This allows us to store the value and
its metadata in a single contiguous block of memory, which requires only
one allocation.

Also, instead of creating our own vtable by storing function pointers
inside the Box, we now use proper virtual function and inheritance. This
is not only more idiomatic C++, but it also reduces Box's size.