Allow GC rooting of arbitrary Traceables
Previously, the only objects that could be rooted were pointers. This is
not perfect from a usability standpoint. For instance, suppose we have
a std::variant of some Ptrs. Such type can't be rooted directly -
instead we have to either allocate it and root the resulting Ptr, or
create a corresponding std::variant with Root members. Both of those
solutions are tedious and inefficient.
A similar problem can be observed for virtually all types which contain
Ptr members and are used on the stack.

To solve this problem, the garbage collector's API is changed to allow
rooting all types that implement Traceable.
Consequently, T in Root<T> now refers to the rooted type, not to the
type behind the pointer. Since Root<Ptr<T>> (rooted pointer) is a very
common use case, we provide a template specialization which should make
working with it a little bit easier.