Simplify the Root API
It turns out that implicit conversions from Root<T> to T are ambiguous
when Root<T> is an rvalue. I don't know how to fix this in an elegant
way, so I just removed the conversions altogether. Additionally, the
template specialization for Root<Ptr<T>> is also removed.
This makes Root's interface a little bit more verbose, but also much
simpler and more explicit.