Make CppFunction an abstract class
We also add helper classes for creating C++ functions and methods
(CppLambda and CppMethod).
Since those classes take a type parameter for a function and lambda
expressions have a type that can't be named, we also add an overload of
Collector::alloc with implicit type inference. Collector::root is also
slightly tweaked to make alloc and root symmetric API-wise.