Add noexcept to Ptr and RootBase copy/move ctors
It turns out that some stdlib functions like std::vector::reserve and
others will not use a move constructor unless it has a noexcept
specifier. Instead, a copy constructor will be used, resulting in
unnecessary object copies.
Root<T>'s implicitly declared move constructor is also affected - it
will not be declared noexcept unless RootBase's copy constructor is
noexcept as well.

To fix those problems, we add noexcept specifiers to copy and move
constructors in Ptr and RootBase. We also do the same for default
constructor in Ptr and assignment operators (because why not).