Fix a memory leak in the GC
Calls to Collector::alloc can be nested. This happens when constructor
for the type being allocated also calls alloc and means that
constructing a type can mutate the linked list of boxes.
Previous implementation didn't take that mutation into account, which
caused values from nested calls to be silently removed from the list.
That in turn caused very mysterious and incomprehensible bugs.

For example, the subclassing Klass constructor allocates an implicit
metaclass. Since the metaclass is allocated by a nested alloc call, it
wouldn't be added to the box list. After that, the first gc cycle would
mark the metaclass, but the sweep phase would never reach and unmark it.
In the second cycle, metaclass is already marked therefore its methods
are never traced, which caused them to be destroyed by the sweep phase.
At this point, metaclass' methods are dangling pointers and accessing
them causes an exception. On top of that, the metaclass itself will
never be deallocated, causing a memory leak.

This should hopefully be fixed now.